# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CampusCoin API
  version: 0.1.0
  description: CampusCoin ("Smart Spending, Student Style") REST API. Money amounts are decimal strings (never floats); errors follow RFC 9457 `application/problem+json`; every route scoped to the caller's own data returns 404 (not 403) for another user's resource.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Short-lived access token (EdDSA JWT) from `POST /auth/login`, sent as `Authorization: Bearer <token>`."
  schemas:
...
    emailOnlySchema:
      type: object
      properties:
        email:
          type: string
          maxLength: 254
      required:
        - email
      additionalProperties: false
...
    loginSchema:
      type: object
      properties:
        email:
          type: string
          maxLength: 254
        password:
          type: string
          minLength: 1
          maxLength: 128
        rememberMe:
          type: boolean
          default: false
      required:
        - email
        - password
      additionalProperties: false
...
    registerSchema:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          maxLength: 254
        password:
          type: string
          minLength: 10
          maxLength: 128
      required:
        - fullName
        - email
        - password
      additionalProperties: false
...
    resetPasswordSchema:
      type: object
      properties:
        token:
          type: string
          pattern: ^[A-Za-z0-9_-]{43}$
        newPassword:
          type: string
          minLength: 10
          maxLength: 128
      required:
        - token
        - newPassword
      additionalProperties: false
...
    verifyEmailSchema:
      type: object
      properties:
        token:
          type: string
          pattern: ^[A-Za-z0-9_-]{43}$
      required:
        - token
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          description: Machine-readable error slug, e.g. "validation-failed".
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
            required:
              - field
              - message
          description: Present on 422 validation-failed responses, one entry per invalid field.
      required:
        - type
        - title
        - status
      description: RFC 9457 application/problem+json body.
...
paths:
  /api/v1/auth/register:
    post:
      tags:
        - Auth
      summary: Register a new student account.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/registerSchema"
      responses:
        "201":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/auth/verify-email:
    post:
      tags:
        - Auth
      summary: Verify email with a one-time token.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/verifyEmailSchema"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/auth/resend-verification:
    post:
      tags:
        - Auth
      summary: Resend the verification email.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/emailOnlySchema"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Request a password-reset email.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/emailOnlySchema"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/auth/reset-password:
    post:
      tags:
        - Auth
      summary: Reset password with a one-time token.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/resetPasswordSchema"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: Log in with email + password; sets the refresh cookie.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/loginSchema"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/auth/refresh:
    post:
      tags:
        - Auth
      summary: Rotate the refresh cookie for a new access token.
      security: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/auth/logout:
    post:
      tags:
        - Auth
      summary: Revoke the current session (refresh-token family).
      security: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/auth/logout-all:
    post:
      tags:
        - Auth
      summary: Revoke every session of the current user.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /api/v1/admin/users:
    get:
      tags:
        - Admin
      summary: Search/list student accounts (admin).
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            minLength: 1
            maxLength: 254
          required: false
          name: q
          in: query
        - schema:
            type: string
            enum:
              - pending
              - active
              - disabled
          required: false
          name: status
          in: query
        - schema:
            type: integer
            minimum: 1
            default: 1
          required: false
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
          name: limit
          in: query
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Authenticated but not allowed (wrong role).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/admin/users/{id}:
    get:
      tags:
        - Admin
      summary: Get one student account (admin).
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Authenticated but not allowed (wrong role).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/admin/users/{id}/disable:
    post:
      tags:
        - Admin
      summary: Disable a student account (admin).
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Authenticated but not allowed (wrong role).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/admin/users/{id}/enable:
    post:
      tags:
        - Admin
      summary: Re-enable a student account (admin).
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Authenticated but not allowed (wrong role).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/admin/users/{id}/send-reset:
    post:
      tags:
        - Admin
      summary: Send a password-reset email to a student (admin).
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Authenticated but not allowed (wrong role).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...